refactor(agent-playground): collapse three shallow seams
* Merge queries/mcp.ts into queries/mcp-queries.ts. The split was
  by CRUD verb (queries vs mutations) for the same domain — inconsistent
  with link-credentials, link-wiring, skill-queries, workspace-queries
  which all co-locate. mcp.ts even imported mcpQueries from mcp-queries.ts
  for invalidation, confirming the seam wasn't earning its keep.

* Delete server/lib/user-agents.ts. The 24-line module wrapped
  UserAdapter with no added behavior — no caching, no error mapping.
  Inline UserAdapter construction at the two callers; execute.ts now
  shares one instance for .exists() and .loadAgent().

* Rename execution-context.svelte.ts -> types/execution-report.ts.
  The .svelte.ts extension was misleading (no runes), and the name
  collided with PlaygroundContextAdapter (the actual execution context).
  File contents are just FSM execution-report types.
