+++ tools/agent-playground/src/lib/types/execution-report.ts
